Incomplete ReadMe
I will soon write a good Readme which explains everything. I will write this and attach the screenshots that I took today after May 1st. I have two midterms this week, but regardless of my midterms I attended DVC Hacks and built this project.  This is a product of the courses that I have completed on Coursera and YouTube.